#29 添加小修改

+++ src/components/UpLoad/index.vue
@@ -1,13 +1,13 @@
 <template>
-  <div class="upload">
-    <el-row class="material" justify type="flex">
-      <el-col>
-        <el-upload :action="materialUploadUrl" :headers="headData" :on-preview="handleMaterialPreview" :on-remove="handleMaterialRemove" :before-remove="MaterialbeforeRemove" multiple :limit="limitFlieNumber" :on-success="handleMaterialSuccess" :file-list="materialfileList" :before-upload="beforeMaterialUpload" :on-exceed="handleMaterialExceed" :on-progress="handleProgress">
-          <el-button size="small" type="primary">点击上传</el-button>
-        </el-upload>
-      </el-col>
-    </el-row>
-  </div>
+    <div class="upload">
+        <el-row class="material" justify type="flex">
+            <el-col>
+                <el-upload :action="materialUploadUrl" :headers="headData" :on-preview="handleMaterialPreview" :on-remove="handleMaterialRemove" :before-remove="MaterialbeforeRemove" multiple :limit="limitFlieNumber" :on-success="handleMaterialSuccess" :file-list="materialfileList" :before-upload="beforeMaterialUpload" :on-exceed="handleMaterialExceed" :on-progress="handleProgress">
+                    <el-button size="small" type="primary">点击上传</el-button>
+                </el-upload>
+            </el-col>
+        </el-row>
+    </div>
 </template>
 
 <script>
@@ -60,20 +60,24 @@ export default {
         },
         handleMaterialRemove(file, fileList) {
             let ids = [];
-            ids.push(file.response.result);
-            this.$emit("update:buttonFlag", true);
-            deleteFile(ids)
-                .then(res => {
-                    if (res.success) {
-                        this.$emit("update:materialfileList", fileList);
+            if (file.response) {
+                ids.push(file.response.result);
+                this.$emit("update:buttonFlag", true);
+                deleteFile(ids)
+                    .then(res => {
+                        if (res.success) {
+                            this.$emit("update:materialfileList", fileList);
+                            this.$emit("update:buttonFlag", false);
+                        } else {
+                            this.$emit("update:buttonFlag", false);
+                        }
+                    })
+                    .catch(() => {
                         this.$emit("update:buttonFlag", false);
-                    } else {
-                        this.$emit("update:buttonFlag", false);
-                    }
-                })
-                .catch(() => {
-                    this.$emit("update:buttonFlag", false);
-                });
+                    });
+            } else {
+                this.$emit("update:buttonFlag", false);
+            }
         },
         handleMaterialPreview(file) {
             // console.log(file);
@@ -86,6 +90,7 @@ export default {
             );
         },
         MaterialbeforeRemove(file, fileList) {
+            console.log("asdasd");
             // return this.$confirm(`确定移除 ${file.name}？`);
             // 文件移除的时候触发的方法
         },
